Update last_biz.pl
Added extra check for the spacer.
When n email arrives with the date in its subject, if the date is between 1-9, the date field will introduce an extra space in front of the $lbd->day and $cbd->day fields. And the regex check which I have commented out will fail since the mail subject it looks for has changed.

My fix uses a spacer variable instead of static space string within if blocks.